// src/Components/Views/Textarea/Textareas.types.ts
interface TextAreasTypes {
    value?: any;
    width?: string;
    placeholder?: string;
    disabled?: boolean;
    minRows?: number | string;
    maxRows?: number | string;
    textColor?: string;
    backgroundColor?: string;
    borderColor?: string;
    focusedBorderColor?: string;
    resize?: "none";
    onChange?: any;
    name?: string;
}

export default TextAreasTypes;